requirements-dev.txt
-r requirements.txt
pytest==7.1.2
black==22.3.0
ruff
coverage==6.4.1
pytest-cov==3.0.0
requests-mock==1.9.3
pytest-flask-sqlalchemy==1.1.0
pre-commit
